+++ src/test/java/iudx/resource/server/apiserver/integrationTests/temporalCountAPIs/GetAfterTemporalCountIT.java
@@ -1,6 +1,6 @@
-package iudx.resource.server.apiserver.integrationtests.temporalSearchAPIs;
+package iudx.resource.server.apiserver.integrationTests.temporalCountAPIs;
 
-import iudx.resource.server.apiserver.integrationtests.RestAssuredConfiguration;
+import iudx.resource.server.apiserver.integrationTests.RestAssuredConfiguration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -10,18 +10,19 @@ import static io.restassured.RestAssured.given;
 import static iudx.resource.server.authenticator.TokensForITs.openResourceToken;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
 
 @ExtendWith(RestAssuredConfiguration.class)
-public class GetBetweenTemporalEntitiesIT {
+public class GetAfterTemporalCountIT {
     String temporalId="b58da193-23d9-43eb-b98a-a103d4b6103c";
     @Test
-    @DisplayName("200 (success) temporal (between)")
-    public void getTemporalEntityTest(){
+    @DisplayName("200 (success) temporal (after)")
+    public void getTemporalCountTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-10-18T14:20:01Z")
+                .queryParam("options","count")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -31,50 +32,32 @@ public class GetBetweenTemporalEntitiesIT {
                 //.log().body()
                 .body("title", equalTo("Success"))
                 .body("type", equalTo("urn:dx:rs:success"))
-                .body("results[0].id", notNullValue());
+                .body("results[0]", notNullValue())
+                .body("results[0].totalHits", is(notNullValue()));
     }
     @Test
-    @DisplayName("200 (success) temporal (between) with optional encryption")
-    public void getTemporalEntityWithOptionalEncryptionTest(){
+    @DisplayName("204 (Empty Response) temporal (after)")
+    public void getTemporalCountWithEmptyResponseTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
-                .header("Content-Type", "application/json")
-                .header("token", openResourceToken)
-                .when()
-                .get("/temporal/entities")
-                .then()
-                .statusCode(200)
-               // .log().body()
-                .body("title", equalTo("Success"))
-                .body("type", equalTo("urn:dx:rs:success"))
-                .body("results[0].id", notNullValue());
-    }
-    @Test
-    @DisplayName("204 (Empty Response) temporal (between)")
-    public void getTemporalEntityWithEmptyResponseTest(){
-        given()
-                .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-01-18T14:20:00Z")
-                .queryParam("endtime", "2020-01-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-01-19T14:20:00Z")
+                .queryParam("options","count")
                 .header("token", openResourceToken)
                 .when()
                 .get("/temporal/entities")
                 .then()
                 .statusCode(204);
-                //.log().body()
+                //.log().body();
     }
     @Test
-    @DisplayName("400 (invalid params) temporal (between)")
-    public void getTemporalEntityWithInvalidParamsTest(){
+    @DisplayName("400 (Invalid params) temporal (after)")
+    public void getTemporalCountWithInvalidParamsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerelation", "during")
-                .queryParam("time", "2020-09-18T14:20:00Z")
-                .queryParam("endtime", "2020-09-19T14:20:00Z")
+                .queryParam("timerelation", "after")
+                .queryParam("timea", "2020-06-01T14:20:01Z")
+                .queryParam("options","count")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -86,13 +69,13 @@ public class GetBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidTemporalRelationParam"));
     }
     @Test
-    @DisplayName("400 (invalid date format) temporal (between)")
-    public void getTemporalEntityWithInvalidDateFormatTest(){
+    @DisplayName("400 (invalid date format) temporal (after)")
+    public void getTemporalCountWithInvalidDateFormatTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-09-18X14:20:00Z")
-                .queryParam("endtime", "2020-09-19X14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01X14:20:01Z")
+                .queryParam("options","count")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -104,14 +87,14 @@ public class GetBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
     }
     @Test
-    @DisplayName("404 (not found) temporal (between)")
-    public void TemporalEntityNotFoundTest(){
-        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b61030";
+    @DisplayName("404 (not found) temporal (after)")
+    public void TemporalCountNotFoundTest(){
+        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b6102c";
         given()
                 .queryParam("id",nonExistingTemporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-09-18T14:20:00Z")
-                .queryParam("endtime", "2020-09-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01T14:20:01Z")
+                .queryParam("options","count")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -123,13 +106,13 @@ public class GetBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:resourceNotFound"));
     }
     @Test
-    @DisplayName("401(invalid credentials) temporal (between)")
-    public void getTemporalEntityWithInvalidCredentialsTest(){
+    @DisplayName("401(invalid credentials) temporal (after)")
+    public void getTemporalCountWithInvalidCredentialsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01T14:20:01Z")
+                .queryParam("options","count")
                 .header("Content-Type", "application/json")
                 .header("token", "abc")
                 .when()
@@ -140,6 +123,4 @@ public class GetBetweenTemporalEntitiesIT {
                 .body("title", equalTo("Not Authorized"))
                 .body("type", equalTo("urn:dx:rs:invalidAuthorizationToken"));
     }
-
-
 }
