+++ src/test/java/iudx/resource/server/apiserver/integrationTests/temporalSearchAPIs/GetBetweenTemporalEntitiesIT.java
@@ -1,6 +1,6 @@
-package iudx.resource.server.apiserver.integrationtests.temporalSearchAPIs;
+package iudx.resource.server.apiserver.integrationTests.temporalSearchAPIs;
 
-import iudx.resource.server.apiserver.integrationtests.RestAssuredConfiguration;
+import iudx.resource.server.apiserver.integrationTests.RestAssuredConfiguration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,15 +12,16 @@ import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
 
 @ExtendWith(RestAssuredConfiguration.class)
-public class GetBeforeTemporalEntitiesIT {
+public class GetBetweenTemporalEntitiesIT {
     String temporalId="b58da193-23d9-43eb-b98a-a103d4b6103c";
     @Test
-    @DisplayName("200 (success) temporal (before)")
+    @DisplayName("200 (success) temporal (between)")
     public void getTemporalEntityTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-10-19T12:00:00Z")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -33,12 +34,32 @@ public class GetBeforeTemporalEntitiesIT {
                 .body("results[0].id", notNullValue());
     }
     @Test
-    @DisplayName("204 (Empty Response) temporal (before)")
+    @DisplayName("200 (success) temporal (between) with optional encryption")
+    public void getTemporalEntityWithOptionalEncryptionTest(){
+        given()
+                .queryParam("id",temporalId)
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .header("Content-Type", "application/json")
+                .header("token", openResourceToken)
+                .when()
+                .get("/temporal/entities")
+                .then()
+                .statusCode(200)
+               // .log().body()
+                .body("title", equalTo("Success"))
+                .body("type", equalTo("urn:dx:rs:success"))
+                .body("results[0].id", notNullValue());
+    }
+    @Test
+    @DisplayName("204 (Empty Response) temporal (between)")
     public void getTemporalEntityWithEmptyResponseTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-01-19T14:20:00Z")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-01-18T14:20:00Z")
+                .queryParam("endtime", "2020-01-19T14:20:00Z")
                 .header("token", openResourceToken)
                 .when()
                 .get("/temporal/entities")
@@ -47,12 +68,13 @@ public class GetBeforeTemporalEntitiesIT {
                 //.log().body()
     }
     @Test
-    @DisplayName("400 (Invalid params) temporal (before)")
+    @DisplayName("400 (invalid params) temporal (between)")
     public void getTemporalEntityWithInvalidParamsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("timea", "2020-09-19T14:20:00Z")
+                .queryParam("timerelation", "during")
+                .queryParam("time", "2020-09-18T14:20:00Z")
+                .queryParam("endtime", "2020-09-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -61,15 +83,16 @@ public class GetBeforeTemporalEntitiesIT {
                 .statusCode(400)
                 //.log().body()
                 .body("title", equalTo("Bad Request"))
-                .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
+                .body("type", equalTo("urn:dx:rs:invalidTemporalRelationParam"));
     }
     @Test
-    @DisplayName("400 (invalid date format) temporal (before)")
+    @DisplayName("400 (invalid date format) temporal (between)")
     public void getTemporalEntityWithInvalidDateFormatTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-09-19X14:20:00Z")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-09-18X14:20:00Z")
+                .queryParam("endtime", "2020-09-19X14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -81,13 +104,14 @@ public class GetBeforeTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
     }
     @Test
-    @DisplayName("404 (not found) temporal (before)")
-    public void temporalEntityNotFoundTest(){
-        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b6102c";
+    @DisplayName("404 (not found) temporal (between)")
+    public void TemporalEntityNotFoundTest(){
+        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b61030";
         given()
                 .queryParam("id",nonExistingTemporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-09-19T14:20:00Z")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-09-18T14:20:00Z")
+                .queryParam("endtime", "2020-09-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -99,12 +123,13 @@ public class GetBeforeTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:resourceNotFound"));
     }
     @Test
-    @DisplayName("401(invalid credentials) temporal (before)")
+    @DisplayName("401(invalid credentials) temporal (between)")
     public void getTemporalEntityWithInvalidCredentialsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-10-19T14:20:00Z")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", "abc")
                 .when()
@@ -115,4 +140,6 @@ public class GetBeforeTemporalEntitiesIT {
                 .body("title", equalTo("Not Authorized"))
                 .body("type", equalTo("urn:dx:rs:invalidAuthorizationToken"));
     }
+
+
 }
